Refactor UserActivity monitor and provider logic
Moved UserActivity+Monitor and Provider implementations from Core/Monitor to Core/Activity, consolidating and simplifying the code. Removed extensive comments and streamlined the structure for better maintainability.
